GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - Virtual branch (refs/gitbutler/Virtual-branch)
   - pasteboard/pasteboard_darwin.go
   - pasteboard/pasteboard.go
   - pasteboard/pastboard_darwin.m
   - bods.go
   - main.go
   - types.go
   - config.go

Your previous branch was: refs/heads/main

The sha for that commit was: 08431433b0ebcbde98f90ea2e030ff4c16866124

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch